+++ core-ng/src/main/java/core/framework/impl/web/site/StaticDirectoryController.java
@@ -1,6 +1,7 @@
 package core.framework.impl.web.site;
 
 import core.framework.http.ContentType;
+import core.framework.http.HTTPHeaders;
 import core.framework.web.Controller;
 import core.framework.web.Request;
 import core.framework.web.Response;
@@ -11,6 +12,7 @@ import org.slf4j.LoggerFactory;
 import java.nio.file.Files;
 import java.nio.file.LinkOption;
 import java.nio.file.Path;
+import java.time.Duration;
 
 /**
  * @author neo
@@ -18,13 +20,15 @@ import java.nio.file.Path;
 public final class StaticDirectoryController implements Controller {
     private final Logger logger = LoggerFactory.getLogger(StaticDirectoryController.class);
     private final Path contentDirectory;
+    private String cacheHeader;
 
-    public StaticDirectoryController(Path contentDirectory) {
+    public StaticDirectoryController(Path contentDirectory, Duration cacheMaxAge) {
         this.contentDirectory = contentDirectory;
+        if (cacheMaxAge != null && cacheMaxAge.getSeconds() > 0) cacheHeader = "public, max-age=" + cacheMaxAge.getSeconds();
     }
 
     @Override
-    public Response execute(Request request) throws Exception {
+    public Response execute(Request request) {
         String path = request.pathParam("path");
         Path filePath = contentDirectory.resolve(path);
         logger.debug("requestFile={}", filePath);
@@ -35,6 +39,7 @@ public final class StaticDirectoryController implements Controller {
         Response response = Response.file(filePath);
         ContentType contentType = MimeTypes.get(filePath.getFileName().toString());
         if (contentType != null) response.contentType(contentType);
+        if (cacheHeader != null) response.header(HTTPHeaders.CACHE_CONTROL, cacheHeader);
         return response;
     }
 }